Fix MSVC C4244 warnings: explicit casts for double->int, float->uint32_t, intptr_t->double
- closest.c: (int)conv_double_to_int_fast() in verify_rounding (will be removed later)
- sparse_image.c: (uint32_t)cut in tosparse_u32, matching tosparse_u16 pattern
- connectedpixels.c: (double)(nf+1), (double)(ns+1) in blob moments init

+++ lib/src/geometry/closest.c
@@ -124,15 +124,15 @@ int verify_rounding(int n) {
 
     for (i = -100; i < 200; i++) {
         v = n + i * 50.0;
-        hfast = conv_double_to_int_fast(v);
-        hslow = conv_double_to_int_safe(v);
+        hfast = (int)conv_double_to_int_fast(v);
+        hslow = (int)conv_double_to_int_safe(v);
         if (hfast != hslow)
             bad++;
     }
     for (i = -100; i < 200; i++) {
         v = -n + i * 50.0;
-        hfast = conv_double_to_int_fast(v);
-        hslow = conv_double_to_int_safe(v);
+        hfast = (int)conv_double_to_int_fast(v);
+        hslow = (int)conv_double_to_int_safe(v);
         if (hfast != hslow)
             bad++;
     }
